// needrelay-frontend/src/api/generated/models/organizerResponseRole.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */

export type OrganizerResponseRole = typeof OrganizerResponseRole[keyof typeof OrganizerResponseRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganizerResponseRole = {
  ADMIN: 'ADMIN',
  ORGANIZER: 'ORGANIZER',
} as const;